# docs/swagger.yaml
basePath: /v1
definitions:
...
  response.ErrorMessage:
    properties:
      error:
        example: Failed to parse UUID
        type: string
    type: object
...
  response.ServerErrorMessage:
    properties:
      error:
        example: Internal server error
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: yu.shuaiwei@tum.de
    name: Shuaiwei Yu
    url: https://github.com/Mtze
  description: 'Measurement instrument for comparing CI job-execution variants. Jobs
    are submitted over REST and completions are observed over REST: the system under
    test posts a terminal status callback to /v1/callback and the benchmarker stamps
    arrival on its own clock. Analysis is done on the raw rows from /v1/export/jobs,
    never on the aggregate endpoints.'
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/Hades-Scheduler/CI-Benchmarker
  title: CI-Benchmarker API
  version: "2.0"
paths:
  /benchmark/build_time/histogram:
    get:
      description: Returns a PNG histogram showing distribution of build time (seconds).
      parameters:
      - description: Start time (RFC3339)
        in: query
        name: from
        type: string
      - description: End time (RFC3339)
        in: query
        name: to
        type: string
      - description: Optional commit hash filter
        in: query
        name: commit_hash
        type: string
      - description: executor filter
        in: query
        name: executor
        required: true
        type: string
      produces:
      - image/png
      responses:
        "200":
          description: PNG image
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ServerErrorMessage'
      summary: Histogram of build time
      tags:
      - metrics
...
  /benchmark/latency/histogram:
    get:
      description: Returns a PNG histogram showing distribution of total latency (seconds).
      parameters:
      - description: Start time (RFC3339)
        in: query
        name: from
        type: string
      - description: End time (RFC3339)
        in: query
        name: to
        type: string
      - description: Optional commit hash filter
        in: query
        name: commit_hash
        type: string
      - description: executor filter
        in: query
        name: executor
        required: true
        type: string
      produces:
      - image/png
      responses:
        "200":
          description: PNG image
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ServerErrorMessage'
      summary: Histogram of total latency
      tags:
      - metrics
...
  /benchmark/queue_latency/histogram:
    get:
      description: Returns a PNG histogram showing distribution of queue latency (seconds).
      parameters:
      - description: Start time (RFC3339)
        in: query
        name: from
        type: string
      - description: End time (RFC3339)
        in: query
        name: to
        type: string
      - description: Optional commit hash filter
        in: query
        name: commit_hash
        type: string
      - description: executor filter
        in: query
        name: executor
        required: true
        type: string
      produces:
      - image/png
      responses:
        "200":
          description: PNG image
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ServerErrorMessage'
      summary: Histogram of queue latency
      tags:
      - metrics
...
schemes:
- http
- https
swagger: "2.0"